Changed name of BaseLogic to TransactionLogic (& Handler)
